+++ src/test/scala/io/github/mkotsur/ProxyLambdaTest.scala
@@ -9,7 +9,7 @@ import io.circe.parser._
 import io.github.mkotsur.ProxyLambdaTest._
 import io.github.mkotsur.aws.handler.Lambda
 import io.github.mkotsur.aws.handler.Lambda._
-import io.github.mkotsur.aws.proxy.{ApiProxyRequest, ProxyRequest, ProxyResponse, RequestContext}
+import io.github.mkotsur.aws.proxy.{ApiProxyRequest, ApiProxyResponse, RequestContext}
 import org.scalatest.concurrent.Eventually
 import org.mockito.MockitoSugar
 import org.scalatest.matchers.should
@@ -22,13 +22,13 @@ object ProxyLambdaTest {
 
   class ProxyRawHandler extends Lambda.ApiProxy[String, RequestContext, String] {
     override protected def handle(input: ApiProxyRequest[String, RequestContext]) =
-      Right(ProxyResponse(200, None, input.body.map(_.toUpperCase())))
+      Right(ApiProxyResponse(200, None, input.body.map(_.toUpperCase())))
   }
 
   class ProxyRawHandlerWithError extends Lambda.ApiProxy[String, RequestContext, String] {
 
     override protected def handle(
-        i: ApiProxyRequest[String, RequestContext]): Either[Throwable, ProxyResponse[String]] =
+        i: ApiProxyRequest[String, RequestContext]): Either[Throwable, ApiProxyResponse[String]] =
       Left(
         new Error("Could not handle this request for some obscure reasons")
       )
@@ -36,7 +36,7 @@ object ProxyLambdaTest {
 
   class ProxyCaseClassHandler extends Lambda.ApiProxy[Ping, RequestContext, Pong] {
     override protected def handle(input: ApiProxyRequest[Ping, RequestContext]) = Right(
-      ProxyResponse(200, None, input.body.map { ping =>
+      ApiProxyResponse(200, None, input.body.map { ping =>
         Pong(ping.inputMsg.length.toString)
       })
     )
@@ -93,9 +93,9 @@ class ProxyLambdaTest extends AnyFunSuite with should.Matchers with MockitoSugar
 
     new ProxyRawHandlerWithError().handle(is, os, mock[Context])
 
-    val response = decode[ProxyResponse[String]](os.toString)
+    val response = decode[ApiProxyResponse[String]](os.toString)
     response shouldEqual Right(
-      ProxyResponse(
+      ApiProxyResponse(
         500,
         Some(Map("Content-Type" -> s"text/plain; charset=UTF-8")),
         Some("Could not handle this request for some obscure reasons")
@@ -111,10 +111,10 @@ class ProxyLambdaTest extends AnyFunSuite with should.Matchers with MockitoSugar
 
     new ProxyCaseClassHandlerWithError().handle(is, os, mock[Context])
 
-    val response = decode[ProxyResponse[String]](os.toString)
+    val response = decode[ApiProxyResponse[String]](os.toString)
 
     response shouldEqual Right(
-      ProxyResponse(
+      ApiProxyResponse(
         500,
         Some(Map("Content-Type" -> s"text/plain; charset=UTF-8")),
         Some("Oh boy, something went wrong...")
@@ -133,8 +133,10 @@ class ProxyLambdaTest extends AnyFunSuite with should.Matchers with MockitoSugar
 
     import Lambda.{canDecodeProxyRequest, canEncodeFuture, canEncodeProxyResponse}
 
-    val function: (ApiProxyRequest[Ping, RequestContext], Context) => Either[Throwable, ProxyResponse[Future[Pong]]] =
-      (_: ApiProxyRequest[Ping, RequestContext], _) => Right(ProxyResponse.success(Some(Future.successful(Pong("4")))))
+    val function
+      : (ApiProxyRequest[Ping, RequestContext], Context) => Either[Throwable, ApiProxyResponse[Future[Pong]]] =
+      (_: ApiProxyRequest[Ping, RequestContext], _) =>
+        Right(ApiProxyResponse.success(Some(Future.successful(Pong("4")))))
     Lambda.Proxy.instance(function).handle(is, os, context)
 
     eventually {
@@ -158,15 +160,15 @@ class ProxyLambdaTest extends AnyFunSuite with should.Matchers with MockitoSugar
 
     Lambda.Proxy
       .instance((_: ApiProxyRequest[Ping, RequestContext], _: Context) => {
-        val response = ProxyResponse.success(Some(Future.failed[String](new RuntimeException("Oops"))))
+        val response = ApiProxyResponse.success(Some(Future.failed[String](new RuntimeException("Oops"))))
         Either.right(response)
       })
       .handle(is, os, context)
 
     eventually {
-      val response = decode[ProxyResponse[String]](os.toString)
+      val response = decode[ApiProxyResponse[String]](os.toString)
       response shouldEqual Either.right(
-        ProxyResponse(
+        ApiProxyResponse(
           500,
           Some(Map("Content-Type" -> s"text/plain; charset=UTF-8")),
           Some("Oops")
@@ -188,7 +190,7 @@ class ProxyLambdaTest extends AnyFunSuite with should.Matchers with MockitoSugar
 
     Lambda.Proxy
       .instance[None.type, RequestContext, None.type]((_, _) => {
-        val response = ProxyResponse[None.type](
+        val response = ApiProxyResponse[None.type](
           statusCode = 200,
           body = None
         )
@@ -197,9 +199,9 @@ class ProxyLambdaTest extends AnyFunSuite with should.Matchers with MockitoSugar
       .handle(is, os, context)
 
     eventually {
-      val response = decode[ProxyResponse[None.type]](os.toString)
+      val response = decode[ApiProxyResponse[None.type]](os.toString)
       response shouldEqual Either.right(
-        ProxyResponse(
+        ApiProxyResponse(
           statusCode = 200,
           body = None
         ))